+++ api/result/bool_result.go
@@ -1,0 +1,4 @@
+package result
+
+// BoolResult is the type of result of a request to the Beget.API, when returning only the bool result
+type BoolResult bool