+++ core/host.go
@@ -1,0 +1,6 @@
+package core
+
+const (
+	Host      = "https://api.beget.com/api/" // prefix of api endpoint
+	UserAgent = "beget-go-server"            // name of user agent
+)