+++ docs.go
@@ -1,0 +1,3 @@
+// Package beget
+// no install
+package beget