+++ pkg/beget/core/host.go
@@ -1,6 +1,0 @@
-package core
-
-const (
-	Host      = "https://api.beget.com/api/"
-	UserAgent = "beget-go-server"
-)