+++ pkg/beget/core/info/api/error_code.go
@@ -1,9 +1,0 @@
-package api
-
-type ErrorCode string
-
-const (
-	AuthError        = ErrorCode("AUTH_ERROR")
-	IncorrectRequest = ErrorCode("INCORRECT_REQUEST")
-	NoSuchMethod     = ErrorCode("NO_SUCH_METHOD")
-)