+++ pkg/beget/core/info/api/formats.go
@@ -1,7 +1,0 @@
-package api
-
-type Format string
-
-const (
-	JSON = Format("json")
-)